Move UltraHDR Sample to graphic folder
