+++ samples/graphics/pdf/src/main/assets/sample.pdf
