+++ samples/graphics/ultrahdr/README.md
@@ -1,0 +1,24 @@
+# UltraHDR Samples
+
+This is a catalog of graphics related UltraHDR samples that demonstrate the usage and capabilities of UltraHDR images on the Android Platform
+
+> 🚧 **Work-in-Progress:** We are activtly working on bringing more UltraHDR samples
+
+Check out the UltraHDR HDR specifications at https://developer.android.com/guide/topics/media/hdr-image-format
+
+## License
+```
+Copyright 2023 The Android Open Source Project
+ 
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+```
