+++ samples/graphics/ultrahdr/src/main/assets/gainmaps/lamps.jpg
