+++ samples/graphics/ultrahdr/src/main/assets/gainmaps/train_station_night.jpg
