+++ samples/graphics/ultrahdr/src/main/res/layout/color_mode_controls.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Copyright (C) 2020 The Android Open Source Project
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/ultrahdr_color_mode_current_mode"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:gravity="center_horizontal"
+        android:padding="@dimen/ultrahdr_color_mode_current_mode_padding"
+        tools:text="@string/ultrahdr_color_mode_sdr" />
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+
+        <Button
+            android:id="@+id/ultrahdr_color_mode_sdr"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/ultrahdr_color_mode_current_mode_padding"
+            android:layout_marginEnd="@dimen/ultrahdr_color_mode_current_mode_padding"
+            android:layout_weight="1"
+            android:text="@string/ultrahdr_color_mode_sdr" />
+
+        <Button
+            android:id="@+id/ultrahdr_color_mode_hdr"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/ultrahdr_color_mode_current_mode_padding"
+            android:layout_marginEnd="@dimen/ultrahdr_color_mode_current_mode_padding"
+            android:layout_weight="1"
+            android:text="@string/ultrahdr_color_mode_hdr" />
+
+    </LinearLayout>
+
+</LinearLayout>
