+++ samples/graphics/ultrahdr/src/main/res/values/dimens.xml
