--- samples/media/ultrahdr/README.md
@@ -1,4 +1,0 @@
-# DisplayingUltraHDRSample samples
-
-// TODO: provide minimal instructions
-```
