--- samples/media/ultrahdr/src/main/res/layout/displaying_ultrahdr_simple.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (C) 2020 The Android Open Source Project
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~      http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <com.example.platform.media.ultrahdr.common.ColorModeControls
-        android:id="@+id/displaying_ultrahdr_color_mode_controls"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-
-    <RadioGroup
-        android:id="@+id/displaying_ultrahdr_samples_selection_group"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/ultrahdr_color_mode_current_mode_padding"
-        android:orientation="horizontal">
-
-        <RadioButton
-            android:id="@+id/displaying_ultrahdr_lamps"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/displaying_ultrahdr_simple_sample_lamps" />
-
-        <RadioButton
-            android:id="@+id/displaying_ultrahdr_canyon"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/displaying_ultrahdr_simple_sample_canyon" />
-
-        <RadioButton
-            android:id="@+id/displaying_ultrahdr_train_station"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/displaying_ultrahdr_simple_sample_train_station" />
-    </RadioGroup>
-
-    <ImageView
-        android:id="@+id/displaying_ultrahdr_image_container"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:layout_gravity="center_horizontal" />
-
-</LinearLayout>
