--- samples/media/ultrahdr/src/main/res/values/strings.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright 2023 The Android Open Source Project
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     https://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<resources>
-    <string name="displaying_ultrahdr_simple_sample_lamps">Lamps</string>
-    <string name="displaying_ultrahdr_simple_sample_canyon">Grand Canyon</string>
-    <string name="displaying_ultrahdr_simple_sample_train_station">Train Station</string>
-
-    <string name="ultrahdr_color_mode_srgb">sRGB</string>
-    <string name="ultrahdr_color_mode_p3">Display P3</string>
-    <string name="ultrahdr_color_mode_hdr">HDR</string>
-    <string name="ultrahdr_color_mode_hdr_with_ratio">HDR | SDR/HDR Ratio = %f</string>
-    <string name="ultrahdr_color_mode_unknown">Unknown</string>
-</resources>
